Fix more long-to-int warnings

+++ Source/Effects/Particles.c
@@ -1319,7 +1319,7 @@ void BurnFire(ObjNode *theNode, float x, float y, float z, Boolean doSmoke,
 {
 int		i;
 float	fps = gFramesPerSecondFrac;
-int		particleGroup;
+long	particleGroup;
 uint32_t magicNum;
 NewParticleGroupDefType	groupDef;
 NewParticleDefType	newParticleDef;
@@ -1339,7 +1339,7 @@ OGLPoint3D			p;
 			theNode->SmokeTimer += SMOKE_TIMER;										// reset timer
 
 			particleGroup 	= theNode->SmokeParticleGroup;
-			magicNum 		= theNode->SmokeParticleMagic;
+			magicNum 		= (uint32_t) theNode->SmokeParticleMagic;
 
 			if ((particleGroup == -1) || (!VerifyParticleGroupMagicNum(particleGroup, magicNum)))
 			{
